api, instance: store env from instance on creation, and use it bind
When creating a new instance, we inspect the container after creating
it, and grab:

 - host ports
 - environment variables

Later on, these host ports and environment variables are used when
binding an application to the instance. The host ports are used along
with the address of the host to build the variable
DIAATS_<plan-name>_INSTANCE, and a map of Docker environment variables
is available in the variable DIAATS_<plan-name>_DOCKER_ENVS. Both values
are JSON-encoded (the former is a list and the latter is an object).

An example, using a MySQL image:

	DIAATS_MYSQL_INSTANCE='["192.168.50.4:3306"]'
	DIAATS_MYSQL_DOCKER_ENVS='{"MYSQL_USER":"root","MYSQL_PASSWORD":"root"}'

Closes #4.